# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ngurra Pathways API
  description: |
    Enterprise API for Ngurra Pathways (Gimbi) - A culturally-grounded employment, 
    education & mentorship platform for First Nations futures.

    ## Rate Limiting

    - Standard endpoints: 1000 requests/hour
    - AI endpoints: 50 requests/hour
    - Enterprise tier: Custom limits

    ## Data Sovereignty

    All data is stored in Australia. Members retain ownership of their personal data
    and can export or delete it at any time. See /data-sovereignty endpoints.

  version: 1.0.0
  contact:
    name: Ngurra Pathways Support
    email: support@ngurrapathways.com.au
  license:
    name: Proprietary
    url: https://ngurrapathways.com.au/terms
servers:
  - url: https://api.ngurrapathways.com.au/v1
    description: Production server
  - url: https://staging-api.ngurrapathways.com.au/v1
    description: Staging server
  - url: https://dev-api.ngurrapathways.com.au/v1
    description: Development server
  - url: http://localhost:3333
    description: Local development
...
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      operationId: postAuthRegister
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login user
      operationId: postAuthLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ------------------------------------------------------------------
  # Jobs
  # ------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminKey:
      type: apiKey
      in: header
      name: x-admin-key
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /oauth/token
          scopes:
            jobs:read: Read job listings
            jobs:write: Create and manage job listings
            members:read: Read member profiles (anonymized)
            analytics:read: Read impact analytics
            rap:read: Read RAP compliance data
            courses:read: Read course listings
            mentorship:read: Read mentorship resources
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true

    # ==========================================
    # AUTHENTICATION SCHEMAS
    # ==========================================
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        userType:
          type: string
          enum: [MEMBER, COMPANY, MENTOR, TAFE]
          default: MEMBER
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
            name:
              type: string
            userType:
              type: string
        message:
          type: string

    # ==========================================
    # SOCIAL NETWORKING SCHEMAS
    # ==========================================
...